Show authors in posts view and clean it up.
The key here is moving each post into its own section in the collection view.
This allows the use of the flow layout's header view (for the author's name),
footer view (for a simple divider), and sectionInset (instead of repeating an
inset in each item's constraints and the item layout). Using the footer view in
turn allows the collection view's background color to be set properly so the
navigation bar doesn't look weird.